Use em value, not percentage, for iframe height
Chrome/Chromium and Firefox differ in their interpretation of the "width"
property, set in CSS, on an iframe. When given a percentage for "width",
Chrome interprets this as a percentage of the surrounding container,
whereas Firefox seems to take a percentage of the height of the document
displayed *inside* the iframe.

To work around this, give the height in ems, not in percent, using
a reasonable-looking value of 15em.